+++ traffic_ops/traffic_ops_golang/login/login.go
@@ -120,7 +120,7 @@ func clientCertAuthentication(w http.ResponseWriter, r *http.Request, db *sqlx.D
 	}
 
 	// Perform certificate verification to ensure it is valid against Root CAs
-	err := auth.VerifyClientCertificate(r, cfg.ClientCertAuth.RootCertsDir)
+	err := auth.VerifyClientCertificate(r, cfg.ClientCertAuth.RootCertsDir, cfg.Insecure)
 	if err != nil {
 		log.Warnf("client cert auth: error attempting to verify client provided TLS certificate. err: %s\n", err)
 		return false